Refactor devfinder and devtools
