+++ include/okami/data/logbook.hpp
@@ -1,0 +1,101 @@
+namespace okami
+{
+namespace LogBook
+{
+enum Enum
+{
+    DestroyTheBoulder,
+    GuardianSaplingTrouble,
+    SecretOfHanaValley,
+    ReviveTheGuardianSaplings,
+    SacredTreeReborn,
+    ShinshuAdventure,
+    HeavySleeper,
+    MakeTheTreesBloom,
+    MrOrangesDance,
+    HaveWaterLilyWillTravel,
+    SaplingOfAgataForest,
+    HardWorkingSon,
+    UmeIsLost,
+    TheLostKey,
+    ASonsDetermination,
+    MysteriousWindmill,
+    DivineWindsCessation,
+    CanineWarriorHunt,
+    RemainingWarriors,
+    AnimalHatingCouple,
+    Rejection,
+    SpringNoMore,
+    BlockedPath,
+    UmesLostAgain,
+    KushisDistress,
+    UmeTheMeal,
+    FightWithHayabusa,
+    DiggingRecord,
+    OrbsReunited,
+    DefeatCrimsonHelm,
+    PowerfulBrew,
+    KushisDecision,
+    KushisFeelings,
+    Festival,
+    ToTheCapital,
+    CityCheckpoint,
+    BringingDownTheBridge,
+    CapitalUnderMistySiege,
+    HimikoQueenOfNippon,
+    WateryWants,
+    FishingForTrouble,
+    LivingSword,
+    FoxRods,
+    SunkenTreasure,
+    TightSecurity,
+    ATinyProblem,
+    Unused1,
+    KaguyasLament,
+    HimikosMotive,
+    LandOfTheDragonians,
+    EmissaryFromTheDeep,
+    TheAmazingOrca,
+    OtohimesTest,
+    IntoTheDragon,
+    FoxRods2,
+    WadasRequest,
+    BadFeelingConfirmed,
+    OniIslandsWhereabouts,
+    LoneWolfVsEvil,
+    ShinshusRollingThunder,
+    SubzeroZone,
+    ColdComfort,
+    SquattersOfEzofuji,
+    IcyWepkeer,
+    LikaIsMissing,
+    ForestOfYoshpet,
+    KaiAsGuide,
+    KaiGoesDeep,
+    PoncleVillagePonctan,
+    GateOfMisfortune,
+    OpenWideTheSpiritGate,
+    SpiritGateHo,
+    BeyondTheGate,
+    KamikiACenturyAgo,
+    ToTheMoonCave,
+    EnterIshaku,
+    TakeLikaToEzofuji,
+    Unused2,
+    OkisUnauthorizedAct,
+    RisingYamato,
+    YoichiThePhilosopher,
+    CarpentersProject,
+    WantedPrayerSlips,
+    ASuspiciousEmperor,
+    KaguyasMemories,
+    CatcallTower,
+    ThePowerOfCatwalk,
+    TwinWhirlpools,
+    CutDownThePeach,
+    OrochisAppetizer,
+    NUM_LOGBOOK_ITEMS,
+};
+const char *GetName(unsigned value);
+} // namespace LogBook
+} // namespace okami
