+++ src/library/logbook.cpp
@@ -1,0 +1,110 @@
+#include "okami/data/logbook.hpp"
+
+namespace okami
+{
+namespace LogBook
+{
+static const char *names[NUM_LOGBOOK_ITEMS] = {
+    "Destroy the Boulder!",
+    "Guardian Sapling Trouble",
+    "Secret of Hana Valley",
+    "Revive the Guardian Saplings",
+    "Sacred Tree Reborn",
+    "Shinshu Adventure",
+    "Heavy Sleeper",
+    "Make the Trees Bloom",
+    "Mr. Orange's Dance",
+    "Have Water Lily, Will Travel",
+    "Sapling of Agata Forest",
+    "Hard Working Son",
+    "Ume is Lost",
+    "The Lost Key",
+    "A Son's Determination",
+    "Mysterious Windmill",
+    "Divine Wind's Cessation",
+    "Canine Warrior Hunt",
+    "Remaining Warriors",
+    "Animal Hating Couple",
+    "Rejection",
+    "Spring No More",
+    "Blocked Path",
+    "Ume's Lost Again",
+    "Kushi's Distress",
+    "Ume the Meal",
+    "Fight With Hayabusa",
+    "Digging Record",
+    "Orbs Reunited",
+    "Defeat Crimson Helm!",
+    "Powerful Brew",
+    "Kushi's Decision",
+    "Kushi's Feelings",
+    "Festival",
+    "To the Capital!",
+    "City Checkpoint",
+    "Bringing Down the Bridge",
+    "Capital Under Misty Siege",
+    "Himiko, Queen of Nippon",
+    "Watery Wants",
+    "Fishing for Trouble",
+    "Living Sword",
+    "Fox Rods",
+    "Sunken Treasure",
+    "Tight Security",
+    "A Tiny Problem",
+    "",
+    "Kaguya's Lament",
+    "Himiko's Motive",
+    "Land of the Dragonians",
+    "Emissary From the Deep",
+    "The Amazing Orca",
+    "Otohime's Test",
+    "Into the Dragon",
+    "Fox Rods",
+    "Wada's Request",
+    "Bad Feeling Confirmed",
+    "Oni Island's Whereabouts",
+    "Lone Wolf vs Evil",
+    "Shinshu's Rolling Thunder",
+    "Subzero Zone",
+    "Cold Comfort",
+    "Squatters of Ezofuji",
+    "Icy Wep'keer",
+    "Lika is Missing!",
+    "Forest of Yoshpet",
+    "Kai as Guide",
+    "Kai Goes Deep",
+    "Poncle Village Ponc'tan",
+    "Gate of Misfortune",
+    "Open Wide the Spirit Gate",
+    "Spirit Gate Ho!",
+    "Beyond the Gate",
+    "Kamiki a Century Ago",
+    "To the Moon Cave!",
+    "Enter Ishaku",
+    "Take Lika to Ezofuji",
+    "",
+    "Oki's Unauthorized Act",
+    "Rising Yamato",
+    "Yoichi the Philosopher",
+    "Carpenter's Project",
+    "Wanted: Prayer Slips",
+    "A Suspicious Emperor",
+    "Kaguya's Memories",
+    "Catcall Tower",
+    "The Power of Catwalk",
+    "Twin Whirlpools",
+    "Cut Down the Peach",
+    "Orochi's Appetizer",
+};
+
+const char *GetName(unsigned value)
+{
+    if (value < NUM_LOGBOOK_ITEMS)
+    {
+        return names[value];
+    }
+    return "invalid";
+}
+
+} // namespace LogBook
+} // namespace okami
